pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "big-benchmark"
version = "0.1.0"
requires-python = ">=3.10,<4.0"
dependencies = [
  "modal~=1.1.3",
  "PyYAML~=6.0.2",
  "stopwatch @ git+https://github.com/modal-labs/stopwatch.git",
]
authors = [
    { name = "Jack Cook", email = "hello@jackcook.com" }
]
description = "A tool for running a lot of benchmarks at once on Modal"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE.md"]
keywords = ["machine-learning", "llms", "vllm", "sglang", "tensorrt-llm"]

[project.optional-dependencies]
tests = [
  "datasette~=0.65.1",
  "fastapi[standard]~=0.116.1",
  "pandas~=2.3.2",
  "pytest~=8.4.1",
  "pytest-xdist~=3.8.0",
  "requests~=2.32.5",
  "SQLAlchemy~=2.0.43",
]

[project.scripts]
bb = "big_benchmark.cli:app"

[project.urls]
Homepage = "https://github.com/modal-labs/big-benchmark"
Documentation = "https://github.com/modal-labs/big-benchmark/blob/main/README.md"
Repository = "https://github.com/modal-labs/big-benchmark.git"

[tool.ruff]
line-length = 88
indent-width = 4
exclude = ["build", "dist", "env", ".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
ignore = [
  "C901",
  "D100",
  "D104",
  "D202",
  "D203",
  "D205",
  "D212",
  "FIX002",
  "PD901",
  "PLC0415",
  "PLR0912",
  "PLR0913",
  "PLR0915",
  "S113",
  "TD003",
  "TRY002", # remove this one later
  "UP017", # sglang image uses python 3.10
]
select = [
  "ALL",
]

[tool.ruff.lint.extend-per-file-ignores]
"src/big_benchmark/cli/*.py" = [
  "T201",
]
"tests/**/test_*.py" = [
  "S101",  # Allow assert statements in tests
  "PLR2004",  # Allow constant values in tests
]
